+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [0.11.0] - 2018-01-22
+
 ### Added
 
 - [Godog](https://github.com/DATA-DOG/godog) runner configuration
@@ -208,7 +211,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bootstrap options
 
 
-[Unreleased]: https://github.com/goph/fxt/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/goph/fxt/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/goph/fxt/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/goph/fxt/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/goph/fxt/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/goph/fxt/compare/v0.7.0...v0.8.0